refactor(app): extract picker controller module
