+++ src/core/app/mod.rs
@@ -1,10 +1,9 @@
-use crate::api::models::{fetch_models, sort_models};
 use crate::auth::AuthManager;
 use crate::core::config::Config;
 use crate::core::message::Message;
 use crate::ui::appearance::{detect_preferred_appearance, Appearance};
 use crate::ui::builtin_themes::{find_builtin_theme, theme_spec_from_custom};
-use crate::ui::picker::{PickerItem, PickerState};
+use crate::ui::picker::PickerState;
 use crate::ui::span::SpanKind;
 use crate::ui::theme::Theme;
 use crate::utils::logging::LoggingState;
@@ -17,96 +16,15 @@ use std::time::Instant;
 use tokio_util::sync::CancellationToken;
 use tui_textarea::TextArea;
 
+pub mod picker;
 pub mod ui_state;
 
+#[cfg_attr(not(test), allow(unused_imports))]
+pub use picker::{
+    ModelPickerState, PickerController, PickerData, PickerSession, ProviderPickerState,
+    ThemePickerState,
+};
 pub use ui_state::{FilePrompt, FilePromptKind, UiMode, UiState};
-
-#[derive(Debug, Clone)]
-pub struct ThemePickerState {
-    pub search_filter: String,
-    pub all_items: Vec<PickerItem>,
-    pub before_theme: Option<Theme>,
-    pub before_theme_id: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-pub struct ModelPickerState {
-    pub search_filter: String,
-    pub all_items: Vec<PickerItem>,
-    pub before_model: Option<String>,
-    pub has_dates: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct ProviderPickerState {
-    pub search_filter: String,
-    pub all_items: Vec<PickerItem>,
-    pub before_provider: Option<(String, String)>,
-}
-
-#[derive(Debug, Clone)]
-#[allow(clippy::large_enum_variant)]
-pub enum PickerData {
-    Theme(ThemePickerState),
-    Model(ModelPickerState),
-    Provider(ProviderPickerState),
-}
-
-#[derive(Debug, Clone)]
-pub struct PickerSession {
-    pub mode: PickerMode,
-    pub state: PickerState,
-    pub data: PickerData,
-}
-
-impl PickerSession {
-    fn prefers_alphabetical(&self) -> bool {
-        match (&self.mode, &self.data) {
-            (PickerMode::Theme, _) | (PickerMode::Provider, _) => true,
-            (PickerMode::Model, PickerData::Model(state)) => !state.has_dates,
-            _ => false,
-        }
-    }
-
-    fn default_sort_mode(&self) -> crate::ui::picker::SortMode {
-        if self.prefers_alphabetical() {
-            crate::ui::picker::SortMode::Name
-        } else {
-            crate::ui::picker::SortMode::Date
-        }
-    }
-
-    fn filter_hint_threshold(&self) -> usize {
-        match self.mode {
-            PickerMode::Model => 20,
-            _ => 10,
-        }
-    }
-
-    fn base_title(&self) -> &'static str {
-        match self.mode {
-            PickerMode::Model => "Pick Model",
-            PickerMode::Provider => "Pick Provider",
-            PickerMode::Theme => "Pick Theme",
-        }
-    }
-
-    fn search_filter(&self) -> &String {
-        match &self.data {
-            PickerData::Model(state) => &state.search_filter,
-            PickerData::Theme(state) => &state.search_filter,
-            PickerData::Provider(state) => &state.search_filter,
-        }
-    }
-
-    fn all_items(&self) -> &Vec<PickerItem> {
-        match &self.data {
-            PickerData::Model(state) => &state.all_items,
-            PickerData::Theme(state) => &state.all_items,
-            PickerData::Provider(state) => &state.all_items,
-        }
-    }
-}
 
 pub struct SessionContext {
     pub client: Client,
@@ -123,32 +41,10 @@ pub struct SessionContext {
     pub startup_env_only: bool,
 }
 
-pub struct PickerController {
-    pub picker_session: Option<PickerSession>,
-    pub in_provider_model_transition: bool,
-    pub provider_model_transition_state: Option<(String, String, String, String, String)>,
-    pub startup_requires_provider: bool,
-    pub startup_requires_model: bool,
-    pub startup_multiple_providers_available: bool,
-}
-
 pub struct App {
     pub session: SessionContext,
     pub ui: UiState,
     pub picker: PickerController,
-}
-
-impl PickerController {
-    pub(crate) fn new() -> Self {
-        Self {
-            picker_session: None,
-            in_provider_model_transition: false,
-            provider_model_transition_state: None,
-            startup_requires_provider: false,
-            startup_requires_model: false,
-            startup_multiple_providers_available: false,
-        }
-    }
 }
 
 fn initialize_logging(
@@ -275,112 +171,56 @@ impl App {
     }
 
     pub fn picker_session(&self) -> Option<&PickerSession> {
-        self.picker.picker_session.as_ref()
+        self.picker.session()
     }
 
+    #[cfg_attr(not(test), allow(dead_code))]
     pub fn picker_session_mut(&mut self) -> Option<&mut PickerSession> {
-        self.picker.picker_session.as_mut()
+        self.picker.session_mut()
     }
 
     pub fn current_picker_mode(&self) -> Option<PickerMode> {
-        self.picker
-            .picker_session
-            .as_ref()
-            .map(|session| session.mode)
+        self.picker.current_mode()
     }
 
     pub fn picker_state(&self) -> Option<&PickerState> {
-        self.picker
-            .picker_session
-            .as_ref()
-            .map(|session| &session.state)
+        self.picker.state()
     }
 
     pub fn picker_state_mut(&mut self) -> Option<&mut PickerState> {
-        self.picker
-            .picker_session
-            .as_mut()
-            .map(|session| &mut session.state)
+        self.picker.state_mut()
     }
 
     pub fn theme_picker_state(&self) -> Option<&ThemePickerState> {
-        match self.picker.picker_session.as_ref() {
-            Some(PickerSession {
-                mode: PickerMode::Theme,
-                data: PickerData::Theme(state),
-                ..
-            }) => Some(state),
-            _ => None,
-        }
+        self.picker.theme_state()
     }
 
     pub fn theme_picker_state_mut(&mut self) -> Option<&mut ThemePickerState> {
-        match self.picker.picker_session.as_mut() {
-            Some(PickerSession {
-                mode: PickerMode::Theme,
-                data: PickerData::Theme(state),
-                ..
-            }) => Some(state),
-            _ => None,
-        }
+        self.picker.theme_state_mut()
     }
 
     pub fn model_picker_state(&self) -> Option<&ModelPickerState> {
-        match self.picker.picker_session.as_ref() {
-            Some(PickerSession {
-                mode: PickerMode::Model,
-                data: PickerData::Model(state),
-                ..
-            }) => Some(state),
-            _ => None,
-        }
+        self.picker.model_state()
     }
 
     pub fn model_picker_state_mut(&mut self) -> Option<&mut ModelPickerState> {
-        match self.picker.picker_session.as_mut() {
-            Some(PickerSession {
-                mode: PickerMode::Model,
-                data: PickerData::Model(state),
-                ..
-            }) => Some(state),
-            _ => None,
-        }
+        self.picker.model_state_mut()
     }
 
     pub fn provider_picker_state(&self) -> Option<&ProviderPickerState> {
-        match self.picker.picker_session.as_ref() {
-            Some(PickerSession {
-                mode: PickerMode::Provider,
-                data: PickerData::Provider(state),
-                ..
-            }) => Some(state),
-            _ => None,
-        }
+        self.picker.provider_state()
     }
 
     pub fn provider_picker_state_mut(&mut self) -> Option<&mut ProviderPickerState> {
-        match self.picker.picker_session.as_mut() {
-            Some(PickerSession {
-                mode: PickerMode::Provider,
-                data: PickerData::Provider(state),
-                ..
-            }) => Some(state),
-            _ => None,
-        }
+        self.picker.provider_state_mut()
     }
 
     /// Close any active picker session.
     pub fn close_picker(&mut self) {
-        self.picker.picker_session = None;
+        self.picker.close();
     }
 
     /// Returns true if current picker should use alphabetical sorting (A–Z / Z–A)
-    fn picker_prefers_alphabetical(&self) -> bool {
-        self.picker_session()
-            .map(|session| session.prefers_alphabetical())
-            .unwrap_or(false)
-    }
-
     pub async fn new_with_auth(
         model: String,
         log_file: Option<String>,
@@ -1036,105 +876,7 @@ impl App {
 
     /// Open a theme picker modal with built-in and custom themes
     pub fn open_theme_picker(&mut self) {
-        let cfg = Config::load_test_safe();
-
-        let mut items: Vec<PickerItem> = Vec::new();
-        let default_theme_id = cfg.theme.clone();
-
-        for t in crate::ui::builtin_themes::load_builtin_themes() {
-            let is_default = default_theme_id
-                .as_ref()
-                .map(|dt| dt.eq_ignore_ascii_case(&t.id))
-                .unwrap_or(false);
-            let label = if is_default {
-                format!("{}*", t.display_name)
-            } else {
-                t.display_name.clone()
-            };
-            let metadata = if is_default {
-                Some("Built-in theme (default from config)".to_string())
-            } else {
-                Some("Built-in theme".to_string())
-            };
-            items.push(PickerItem {
-                id: t.id.clone(),
-                label,
-                metadata,
-                sort_key: Some(t.display_name.clone()),
-            });
-        }
-
-        for ct in cfg.list_custom_themes() {
-            let is_default = default_theme_id
-                .as_ref()
-                .map(|dt| dt.eq_ignore_ascii_case(&ct.id))
-                .unwrap_or(false);
-            let base_label = format!("{} (custom)", ct.display_name);
-            let label = if is_default {
-                format!("{}*", base_label)
-            } else {
-                base_label
-            };
-            let metadata = if is_default {
-                Some("Custom theme (config.toml) (default from config)".to_string())
-            } else {
-                Some("Custom theme (config.toml)".to_string())
-            };
-            items.push(PickerItem {
-                id: ct.id.clone(),
-                label,
-                metadata,
-                sort_key: Some(ct.display_name.clone()),
-            });
-        }
-
-        let active_theme_id = self
-            .ui
-            .current_theme_id
-            .as_ref()
-            .or(cfg.theme.as_ref())
-            .cloned();
-
-        let mut selected = 0usize;
-        if let Some(id) = &active_theme_id {
-            if let Some((idx, _)) = items
-                .iter()
-                .enumerate()
-                .find(|(_, it)| it.id.eq_ignore_ascii_case(id))
-            {
-                selected = idx;
-            }
-        }
-
-        let picker_state = PickerState::new("Pick Theme", items.clone(), selected);
-        let mut session = PickerSession {
-            mode: PickerMode::Theme,
-            state: picker_state,
-            data: PickerData::Theme(ThemePickerState {
-                search_filter: String::new(),
-                all_items: items,
-                before_theme: Some(self.ui.theme.clone()),
-                before_theme_id: cfg.theme.clone(),
-            }),
-        };
-
-        session.state.sort_mode = session.default_sort_mode();
-        self.picker.picker_session = Some(session);
-
-        self.sort_picker_items();
-        self.update_picker_title();
-
-        if let (Some(theme_id), Some(session)) = (active_theme_id, self.picker_session_mut()) {
-            if let Some((idx, _)) = session
-                .state
-                .items
-                .iter()
-                .enumerate()
-                .find(|(_, it)| it.id.eq_ignore_ascii_case(&theme_id))
-            {
-                session.state.selected = idx;
-            }
-        }
+        self.picker.open_theme_picker(&mut self.ui);
     }
 
     /// Apply theme by id (custom or built-in) and persist in config
@@ -1201,339 +943,33 @@ impl App {
 
     /// Open a model picker modal with available models from current provider
     pub async fn open_model_picker(&mut self) -> Result<(), Box<dyn std::error::Error>> {
-        let cfg = Config::load_test_safe();
-        let default_model_for_provider =
-            cfg.get_default_model(&self.session.provider_name).cloned();
-
-        let models_response = fetch_models(
-            &self.session.client,
-            &self.session.base_url,
-            &self.session.api_key,
-            &self.session.provider_name,
-        )
-        .await?;
-
-        if models_response.data.is_empty() {
-            return Err("No models available from this provider".into());
-        }
-
-        let mut models = models_response.data;
-        sort_models(&mut models);
-
-        let has_dates = models.iter().any(|m| {
-            m.created.map(|v| v > 0).unwrap_or(false)
-                || m.created_at
-                    .as_ref()
-                    .map(|s| s.len() > 4 && (s.contains('-') || s.contains('/')))
-                    .unwrap_or(false)
-        });
-
-        let items: Vec<PickerItem> = models
-            .into_iter()
-            .map(|model| {
-                let mut label = if let Some(display_name) = &model.display_name {
-                    if display_name != &model.id && !display_name.is_empty() {
-                        format!("{} ({})", model.id, display_name)
-                    } else {
-                        model.id.clone()
-                    }
-                } else {
-                    model.id.clone()
-                };
-
-                if let Some(ref def) = default_model_for_provider {
-                    if def.eq_ignore_ascii_case(&model.id) {
-                        label.push('*');
-                    }
-                }
-
-                let metadata = if let Some(created) = model.created {
-                    if created > 0 && created < u64::MAX / 1000 {
-                        let timestamp_secs = if created > 10_000_000_000 {
-                            created / 1000
-                        } else {
-                            created
-                        };
-
-                        if timestamp_secs > 0 && timestamp_secs < 32_503_680_000 {
-                            chrono::DateTime::<chrono::Utc>::from_timestamp(
-                                timestamp_secs as i64,
-                                0,
-                            )
-                            .map(|datetime| {
-                                format!("Created: {}", datetime.format("%Y-%m-%d %H:%M UTC"))
-                            })
-                        } else {
-                            Some(format!("Created: {} (invalid timestamp)", created))
-                        }
-                    } else {
-                        None
-                    }
-                } else if let Some(created_at) = &model.created_at {
-                    if !created_at.is_empty() {
-                        if created_at.len() > 4
-                            && (created_at.contains('-') || created_at.contains('/'))
-                        {
-                            Some(format!("Created: {}", created_at))
-                        } else {
-                            Some(format!("Created: {} (unrecognized format)", created_at))
-                        }
-                    } else {
-                        None
-                    }
-                } else {
-                    model
-                        .owned_by
-                        .as_ref()
-                        .filter(|owner| !owner.is_empty() && *owner != "system")
-                        .map(|owner| format!("Owner: {}", owner))
-                };
-
-                let sort_key = if has_dates {
-                    model
-                        .created
-                        .filter(|&created| created > 0)
-                        .map(|created| format!("{:020}", created))
-                } else {
-                    None
-                };
-
-                PickerItem {
-                    id: model.id,
-                    label,
-                    metadata,
-                    sort_key,
-                }
-            })
-            .collect();
-
-        let mut selected = 0usize;
-        if let Some((idx, _)) = items
-            .iter()
-            .enumerate()
-            .find(|(_, it)| it.id == self.session.model)
-        {
-            selected = idx;
-        }
-
-        let picker_state = PickerState::new("Pick Model", items.clone(), selected);
-        let mut session = PickerSession {
-            mode: PickerMode::Model,
-            state: picker_state,
-            data: PickerData::Model(ModelPickerState {
-                search_filter: String::new(),
-                all_items: items,
-                before_model: Some(self.session.model.clone()),
-                has_dates,
-            }),
-        };
-
-        session.state.sort_mode = session.default_sort_mode();
-        self.picker.picker_session = Some(session);
-
-        self.sort_picker_items();
-        self.update_picker_title();
-
-        let current_model = self.session.model.clone();
-        if let Some(session) = self.picker_session_mut() {
-            if let Some((idx, _)) = session
-                .state
-                .items
-                .iter()
-                .enumerate()
-                .find(|(_, it)| it.id == current_model)
-            {
-                session.state.selected = idx;
-            }
-        }
-
-        Ok(())
+        self.picker.open_model_picker(&self.session).await
     }
 
     /// Filter models based on search term and update picker
     pub fn filter_models(&mut self) {
-        let Some(session) = self.picker_session_mut() else {
-            return;
-        };
-        if let (PickerMode::Model, PickerData::Model(model_state)) =
-            (session.mode, &mut session.data)
-        {
-            let search_term = model_state.search_filter.to_lowercase();
-            let filtered: Vec<PickerItem> = if search_term.is_empty() {
-                model_state.all_items.clone()
-            } else {
-                model_state
-                    .all_items
-                    .iter()
-                    .filter(|item| {
-                        item.id.to_lowercase().contains(&search_term)
-                            || item.label.to_lowercase().contains(&search_term)
-                    })
-                    .cloned()
-                    .collect()
-            };
-            session.state.items = filtered;
-            if session.state.selected >= session.state.items.len() {
-                session.state.selected = 0;
-            }
-            self.sort_picker_items();
-            self.update_picker_title();
-        }
+        self.picker.filter_models();
     }
 
     /// Filter themes based on search term and update picker
     pub fn filter_themes(&mut self) {
-        let Some(session) = self.picker_session_mut() else {
-            return;
-        };
-        if let (PickerMode::Theme, PickerData::Theme(theme_state)) =
-            (session.mode, &mut session.data)
-        {
-            let search_term = theme_state.search_filter.to_lowercase();
-            let filtered: Vec<PickerItem> = if search_term.is_empty() {
-                theme_state.all_items.clone()
-            } else {
-                theme_state
-                    .all_items
-                    .iter()
-                    .filter(|item| {
-                        item.id.to_lowercase().contains(&search_term)
-                            || item.label.to_lowercase().contains(&search_term)
-                    })
-                    .cloned()
-                    .collect()
-            };
-            session.state.items = filtered;
-            if session.state.selected >= session.state.items.len() {
-                session.state.selected = 0;
-            }
-            self.sort_picker_items();
-            self.update_picker_title();
-        }
+        self.picker.filter_themes();
     }
 
     /// Filter providers based on search term and update picker
     pub fn filter_providers(&mut self) {
-        let Some(session) = self.picker_session_mut() else {
-            return;
-        };
-        if let (PickerMode::Provider, PickerData::Provider(provider_state)) =
-            (session.mode, &mut session.data)
-        {
-            let search_term = provider_state.search_filter.to_lowercase();
-            let filtered: Vec<PickerItem> = if search_term.is_empty() {
-                provider_state.all_items.clone()
-            } else {
-                provider_state
-                    .all_items
-                    .iter()
-                    .filter(|item| {
-                        item.id.to_lowercase().contains(&search_term)
-                            || item.label.to_lowercase().contains(&search_term)
-                    })
-                    .cloned()
-                    .collect()
-            };
-            session.state.items = filtered;
-            if session.state.selected >= session.state.items.len() {
-                session.state.selected = 0;
-            }
-            self.sort_picker_items();
-            self.update_picker_title();
-        }
+        self.picker.filter_providers();
     }
 
     /// Sort picker items based on current sort mode
     pub fn sort_picker_items(&mut self) {
-        let prefers_alpha = self.picker_prefers_alphabetical();
-        if let Some(session) = self.picker_session_mut() {
-            let picker = &mut session.state;
-            if prefers_alpha {
-                // For themes and providers, use alphabetical sorting
-                match picker.sort_mode {
-                    crate::ui::picker::SortMode::Date => {
-                        // Z-A sorting (reverse alphabetical)
-                        picker.items.sort_by(|a, b| b.label.cmp(&a.label));
-                    }
-                    crate::ui::picker::SortMode::Name => {
-                        // A-Z sorting (normal alphabetical)
-                        picker.items.sort_by(|a, b| a.label.cmp(&b.label));
-                    }
-                }
-            } else {
-                // For models, use timestamp-based sorting
-                match picker.sort_mode {
-                    crate::ui::picker::SortMode::Date => {
-                        // Sort by sort_key (timestamp or creation date), newest first
-                        picker.items.sort_by(|a, b| {
-                            match (&a.sort_key, &b.sort_key) {
-                                (Some(a_key), Some(b_key)) => b_key.cmp(a_key), // Reverse for newest first
-                                (Some(_), None) => std::cmp::Ordering::Less,
-                                (None, Some(_)) => std::cmp::Ordering::Greater,
-                                // If both missing, reverse alphabetical to distinguish from Name mode
-                                (None, None) => b.label.cmp(&a.label),
-                            }
-                        });
-                    }
-                    crate::ui::picker::SortMode::Name => {
-                        // Sort by label alphabetically
-                        picker.items.sort_by(|a, b| a.label.cmp(&b.label));
-                    }
-                }
-            }
-
-            // Reset selection to first item after sorting
-            if picker.selected >= picker.items.len() {
-                picker.selected = 0;
-            }
-        }
+        self.picker.sort_items();
     }
 
     /// Update picker title to show sort mode
     pub fn update_picker_title(&mut self) {
-        let Some(session) = self.picker_session_mut() else {
-            return;
-        };
-
-        let prefers_alpha = session.prefers_alphabetical();
-        let base_title = session.base_title();
-        let item_count = session.all_items().len();
-        let threshold = session.filter_hint_threshold();
-        let search_filter = session.search_filter().clone();
-
-        let picker = &mut session.state;
-        let sort_text = if prefers_alpha {
-            match picker.sort_mode {
-                crate::ui::picker::SortMode::Name => "A-Z",
-                crate::ui::picker::SortMode::Date => "Z-A",
-            }
-        } else {
-            match picker.sort_mode {
-                crate::ui::picker::SortMode::Date => "date",
-                crate::ui::picker::SortMode::Name => "name",
-            }
-        };
-
-        picker.title = if search_filter.is_empty() {
-            if item_count > threshold {
-                format!(
-                    "{} ({} available - Sort by: {} - type to filter)",
-                    base_title, item_count, sort_text
-                )
-            } else {
-                format!("{} (Sort by: {})", base_title, sort_text)
-            }
-        } else {
-            format!(
-                "{} (filter: '{}' - {} matches - Sort by: {})",
-                base_title,
-                search_filter,
-                picker.items.len(),
-                sort_text
-            )
-        }
+        self.picker.update_title();
     }
-
     /// Apply model by id and persist in current session (not config)
     pub fn apply_model_by_id(&mut self, model_id: &str) {
         self.session.model = model_id.to_string();
@@ -1570,122 +1006,8 @@ impl App {
 
     /// Open a provider picker modal with available providers
     pub fn open_provider_picker(&mut self) {
-        use crate::auth::AuthManager;
-        use crate::core::builtin_providers::load_builtin_providers;
-
-        let auth_manager = AuthManager::new();
-        let cfg = Config::load_test_safe();
-        let default_provider = cfg.default_provider.clone();
-        let mut items: Vec<PickerItem> = Vec::new();
-
-        // Add built-in providers that have authentication
-        let builtin_providers = load_builtin_providers();
-        for builtin_provider in builtin_providers {
-            if let Ok(Some(_)) = auth_manager.get_token(&builtin_provider.id) {
-                let is_default = default_provider
-                    .as_ref()
-                    .map(|dp| dp.eq_ignore_ascii_case(&builtin_provider.id))
-                    .unwrap_or(false);
-                let label = if is_default {
-                    format!("{}*", builtin_provider.display_name)
-                } else {
-                    builtin_provider.display_name.clone()
-                };
-                let metadata = if is_default {
-                    Some(format!(
-                        "Built-in provider ({}) (default from config)",
-                        builtin_provider.base_url
-                    ))
-                } else {
-                    Some(format!("Built-in provider ({})", builtin_provider.base_url))
-                };
-                items.push(PickerItem {
-                    id: builtin_provider.id.clone(),
-                    label,
-                    metadata,
-                    sort_key: Some(builtin_provider.display_name.clone()),
-                });
-            }
-        }
-
-        // Add custom providers that have authentication
-        let custom_providers = auth_manager.list_custom_providers();
-        for (id, display_name, base_url, has_token) in custom_providers {
-            if has_token {
-                let is_default = default_provider
-                    .as_ref()
-                    .map(|dp| dp.eq_ignore_ascii_case(&id))
-                    .unwrap_or(false);
-                let label = if is_default {
-                    format!("{} (custom)*", display_name)
-                } else {
-                    format!("{} (custom)", display_name)
-                };
-                let metadata = if is_default {
-                    Some(format!(
-                        "Custom provider ({}) (default from config)",
-                        base_url
-                    ))
-                } else {
-                    Some(format!("Custom provider ({})", base_url))
-                };
-                items.push(PickerItem {
-                    id,
-                    label,
-                    metadata,
-                    sort_key: Some(display_name),
-                });
-            }
-        }
-
-        if items.is_empty() {
-            // No providers available, show error
-            self.set_status(
-                "No configured providers found. Run 'chabeau auth' to set up authentication.",
-            );
-            return;
-        }
-
-        let mut selected = 0usize;
-        if let Some((idx, _)) = items
-            .iter()
-            .enumerate()
-            .find(|(_, it)| it.id == self.session.provider_name)
-        {
-            selected = idx;
-        }
-
-        let picker_state = PickerState::new("Pick Provider", items.clone(), selected);
-        let mut session = PickerSession {
-            mode: PickerMode::Provider,
-            state: picker_state,
-            data: PickerData::Provider(ProviderPickerState {
-                search_filter: String::new(),
-                all_items: items,
-                before_provider: Some((
-                    self.session.provider_name.clone(),
-                    self.session.provider_display_name.clone(),
-                )),
-            }),
-        };
-
-        session.state.sort_mode = session.default_sort_mode();
-        self.picker.picker_session = Some(session);
-
-        self.sort_picker_items();
-        self.update_picker_title();
-
-        let current_provider = self.session.provider_name.clone();
-        if let Some(session) = self.picker_session_mut() {
-            if let Some((idx, _)) = session
-                .state
-                .items
-                .iter()
-                .enumerate()
-                .find(|(_, it)| it.id == current_provider)
-            {
-                session.state.selected = idx;
-            }
+        if let Err(message) = self.picker.open_provider_picker(&self.session) {
+            self.set_status(message);
         }
     }
 
@@ -1695,8 +1017,6 @@ impl App {
     /// - Result<(), String> indicates success or failure of the provider change
     /// - bool indicates whether a model picker should be opened after changing provider
     pub fn apply_provider_by_id(&mut self, provider_id: &str) -> (Result<(), String>, bool) {
-        use crate::auth::AuthManager;
-
         let auth_manager = AuthManager::new();
         let config = Config::load_test_safe();
 
@@ -1878,6 +1198,7 @@ const _: fn(Theme, bool, bool) -> App = App::new_bench;
 mod tests {
     use super::*;
     use crate::core::text_wrapping::{TextWrapper, WrapConfig};
+    use crate::ui::picker::PickerItem;
     use crate::utils::test_utils::{create_test_app, create_test_message};
     use tui_textarea::{CursorMove, Input, Key};
